11-7
Today I added the content to the Plan a Visit page and began to stylize that. I also added more content to the about page. 

There is some information that I currently want to rework and possible create a new page to because the about page has a lot of content. I either need to flush it down, create a new page, or design a good functioning and accessible layout to include all of the content.