11-21
Note to self of things needing further attention:
Home Page:
	The arrows for the slick scroll are no longer working and I am not able to find the code that is causing it, unless it is because of creating the placement for the images under the banner and it is causing the interactive error?
	Spacing on the bottom image came back. It was not there right at first but then I added margin top to the footer content. (It didn’t do anything when I put the margin-top on the whole div so I had to do each element separate and it worked) and there is space and it is still there if I delete the margin top. I changed the z-index and that is not doing anything either. I was trying to move the footer up because in the bottom right corner, it is starting to go over the image slightly.

About Page:
The columns are not lining up and I added a padding to style that section however each h1 and p section end up being different sizes and seem to have a different margin space causing different gaps between them. I am not sure.

Animal Page:
The X in the modal after I added the scrolling feature within modal is partially cut off and because of scrolling function can not click on x every time to exit screen

Education:
Broken code on the accordion. I reused the same code from the about park info section and renamed the div centerme_events however I am not finding the code that is causing all the accordions to be in a row instead of a column.